1000以上の戦果が表示されないのを直す

+++ KancolleSniffer/MainForm.cs
@@ -303,7 +303,7 @@ namespace KancolleSniffer
             var ac = _sniffer.Achievement.Value;
             if (ac >= 10000)
                 ac = 9999;
-            labelAchievement.Text = ac >= 1000 ? ac.ToString("D") : ac.ToString("F1");
+            labelAchievement.Text = ac >= 1000 ? ((int)ac).ToString("D") : ac.ToString("F1");
             UpdateMaterialHistry();
             if (_shipListForm.Visible)
                 _shipListForm.UpdateList();
